wizard/storage/requirements.txt
grpcio
grpcio-tools
protobuf
psycopg2-binary  # For Postgres DB connection
cryptography     # For AES-GCM encryption